backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.25.1
tenacity==8.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
prometheus-client==0.19.0
sentry-sdk==1.35.0
python-json-logger==2.0.7
openai==1.3.0
edge-tts==6.1.9
soundfile==0.12.1
numpy==1.24.3
pydub==0.25.1
python-socketio==5.10.0
